+++ Expanded Cheat Menu/Code/Menus/HelpFunc.lua
@@ -352,9 +352,9 @@ do -- ModUpload
 
 						if mod.id == ChoGGi.id then
 							mod.description = Translate(302535920000990--[["You need to have a mouse to use this mod."]]) .. "\n"
-								.. needs_lib .. Translate(302535920000887--[["If you get a disabled content restrictions error: Please let me know and I'll tell Paradox (can take a few days or weeks).
+								.. needs_lib .. Translate(302535920000887--[["Mod disabled by content restrictions: https://steamcommunity.com/sharedfiles/filedetails/?id=2835850170
 
-If you have any issues with this mod then please send me a bug report instead of reporting the mod.
+If you have any issues with this mod then please send me a bug report instead of reporting the mod (or it'll get taken down).
 https://github.com/ChoGGi/SurvivingMars_CheatMods
 Discord: https://discord.gg/ZXXYaExThy
 https://steamcommunity.com/id/ChoGGi/