+++ Mods ChoGGi/ChoGGi Construction Layouts/metadata.lua
@@ -12,16 +12,17 @@ return PlaceObj("ModDef", {
 	"lua_revision", 1007000, -- Picard
 	"steam_id", "2234863273",
 	"pops_any_uuid", "77fd2531-04b6-4e98-9968-f21253455476",
-	"version", 9,
-	"version_major", 0,
-	"version_minor", 9,
+	"version", 10,
+	"version_major", 1,
+	"version_minor", 0,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
 	"TagBuildings", true,
-	"description", [[A bunch of construction layouts:
+	"description", [[
+A bunch of construction layouts:
 
 Repeatable Stirlings + Tribby (Infrastructure).
 Repeatable large Solar Panels (Power).