+++ Mods ChoGGi/Dome Show Work Slots/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "3524880374",
 	"pops_any_uuid", "f713b60f-0f3c-474c-8ff6-732f03b0677a",
 	"lua_revision", 1007000, -- Picard
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {